[Spring 지하철 노선도 - 3단계] 디우(김동규) 미션 제출합니다. (#273)
* feat: 인수테스트 Fixture 적용

RestAssured를 활용하여 REST API를 테스트하는 부분을 Fixture로 분리했습니다.

* feat: id 추출부분을 별도의 메소드로 분리

id를 추출하는 부분을 별도의 메소드로 분리하여 코드의 가독성을 개선해보았습니다.

* refactor: Dao 인터페이스 도입

DAO 의 기능명세를 위한 인터페이스를 도입하였습니다.

* feat: Fake DAO 적용

가짜 DAO객체를 통해서 Service를 테스트하도록 하였습니다.

* refactor: 사용하지 않는 생성자 제거

사용하지 않는 기본생성자를 제거하였습니다.

* refactor: 생성자 추가

Line 도메인을 받아와 초기화를 진행하는 생성자를 추가하였습니다.

* feat: 인자 크기 검증 추가

지정된 컬럼의 크기를 초과하는 값에 대한 처리를 추가해주었습니다.

* docs(README): 지하철 구간 관리 기능 요구사항 정리

* feat: Section 도메인

Section의 두 종점 간의 거리는 0보다 작거나 같을 수 없으며, 상행종점과 하행 종점은 같은 역일 수 없다.

* feat: 구간 등록

새로 등록할 구간의 상행역과 하행역 중 노선에 이미 등록되어 있는 역을 기준으로 새로운 구간을 추가한다.

* feat: 구간 등록

하나의 노선에는 갈래길이 허용되지 않기 때문에 새로운 구간이 추가되기 전에 갈래길이 생기지 않도록 기존 구간을 변경한다.

* feat: 구간 등록

역 사이에 새로운 역을 등록할 경우 기존 역 사이 길이보다 크거나 같으면 등록을 할 수 없다.

* feat: 구간 등록

상행역과 하행역이 이미 노선에 모두 등록되어 있다면 추가할 수 없다.

* feat: 구간 등록

상행역과 하행역 둘 중 하나도 포함되어 있지 않으면 추가할 수 없다.

* feat: 구간 제거

Station을 통해서 구간을 제거할 수 있다.

* feat: 구간 제거

구간이 하나인 노선에서 마지막 구간을 제거할 수 없다.

* feat: SectionDao 구현

* feat: Line에 Sections 추가 및 생성자 추가

* feat: 정렬된 Section 리스트 반환

* feat: response DTO에 Station에 대한 내용 추가

* feat: LineService 기존 테스트 통과하도록 수정

* feat: E2E Line 테스트시 변경된 사항(station정보) 추가하도록 수정

* feat: 단건 노선에 대한 구간 조회 구현

* feat: SectionService 등록 및 삭제 구현

* feat: SectionController 구현

* feat: 불필요한 코드 제거 및 구간 추가시 검사

* test: 환승역(교점)에 대한 구간 테스트

* feat: Sections를 이용하도록 수정

* feat: 구간 등록 테스트 추가

* refactor: 불필요한 생성자 및 필드 제거

* refactor: 불필요한 생성자 및 필드 제거

* refactor: EmptyResultDataAccessException 제거

* fix: 노선 등록시 최초의 구간을 등록하므로 이를 검사하는 로직 제거

* feat: batchUpdate 적용

* fix: 갈래길 방지 로직 수정

하행이 같은 경우에 대해서 처리를 못했었던 문제를 수정하였습니다.

* feat: 노선 삭제시 구간들 함께 삭제

* feat: 역 이전에 구간 먼저 삭제

역을 삭제하기 위해서는 해당 역을 사용중인 구간을 먼저 삭제해야만 삭제 가능하도록 구현하였습니다.

* feat: 처음과 끝 역 삭제 구현

처음과 끝 역에 대해서도 삭제가 가능하도록 로직을 추가해주었습니다.

* refactor: PathVariable 생략

* refactor: 의미를 명확히 하기 위해 메소드 분리

`extractStationIdFromName` 이라는 메소드는 "ExtractableResponse 만들기" 와 "id값 추출" 이라는 두가지 책임을 가지고 있었으므로 이를 `makeStationResponse` 와 `extractStationId` 두 개의 메소드로 분리하여 그 의미를 명확히 하였습니다.

* refactor: AcceptanceFixture 사용

기존에 AcceptanceFixture 재사용이 누락된 부분에 AcceptanceFixture.get() 을 적용하였습니다.

* refactor: DisplayName 을 구체적으로 수정

DisplayName을 수정하여 무엇을 테스트하는 것인지를 보다 명확히 알 수 있도록 하였습니다.

* refactor: 테스트 코드에서의 orElseThrow 제거

orElseThrow를 제거하고 대신 "실패"를 의미하는 객체를 생성하여 orElseGet()으로 null인경우에만 반환하도록 하였습니다.

* feat: 지하철역 이름 검사에 대한 테스트 추가

지하철역이 255를 초과하는 경우에 대한 테스트 코드 추가

* feat: StationName VO 적용

* refactor: 예외 메시지 변경

종점은 노선의 양 끝역을 이야기하므로, 제대로된 예외 메시지가 아니어서 이를 수정해주었습니다.

* refactor: 메소드명 변경

isContainStation은 `is + 동사원형` 형태이므로 `containsStation` 으로 수정하였습니다.

* refactor: final 키워드 추가

* refactor: 불필요한 공백 제거

* refactor: 메소드 분리 및 메소드명 변경

* refactor: 중복되는 부분에 대해서 메소드 추출

* refactor: 직관적인 메소드명으로 변경

* refactor: 예외 메시지를 구체적으로 수정

* refactor: 코드 간략화

스트림을 적용하여 메소드를 제거하고, 변수명 수정

* refactor: @Transactional 어노테이션 추가

@Transactional 어노테이션을 추가하여 하나의 메소드가 하나의 트랙젼션이 되도록 하였습니다.

* refactor: 메소드 순서 정리

* refactor: DisplayName 수정

* refactor: 예외 메시지 수정

* refactor: 불필요한 given when then 주석 제거

* refactor: 파일의 끝에 개행 추가

* refactor: DisplayName 구체적으로 수정

* refactor: Section 식별자 제거

Section의 식별자가 불필요한 곳에서 제거해주었습니다.

* refactor: INAPPROPRIATE_LINE 추가

Optional이 비어있을 미리 생성해둔 상수(static final)에 해당하는 INAPPROPRIATE_LINE을 반환

* refactor: Sections 정렬된 상태 유지

Sections가 생성되는 시점 그리고 데이터 변경이 일어나는 add, delete 순간에도 정렬을 수행하도록 수정하였습니다.

* refactor: size검사와 내용 검사 하나로 체이닝

* refactor: 부적절한 isXXX 수정 및 제거

* refactor: 예외 메시지 변경

* refactor: 같은 Station 비교 구현

값을 직접 가져와서 비교하는게 아니라 Section에 같은 Station인지 비교하는 메소드를 두어, 물어보도록 수정하였습니다.

* refactor: validateThatHasMinimumSection -> validateDeletableSize 네이밍 변경

* refactor: 구간 삭제 로직 수정

Sections가 항상 정렬된 상태를 유지하므로 종점 삭제시 index 0과 size()-1 을 이용함으로써 구간을 찾는 로직을 제거해주었습니다.

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -2,23 +2,31 @@ package wooteco.subway.service;
 
 import static java.util.stream.Collectors.toUnmodifiableList;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
+import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
 import wooteco.subway.exception.NotExistException;
 
 @Service
+@Transactional
 public class StationService {
 
     private static final int DELETE_FAIL = 0;
 
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse save(StationRequest request) {
@@ -29,6 +37,7 @@ public class StationService {
         return new StationResponse(savedId, station.getName());
     }
 
+    @Transactional(readOnly = true)
     public List<StationResponse> findAll() {
         return stationDao.findAll()
                 .stream()
@@ -37,10 +46,31 @@ public class StationService {
     }
 
     public void deleteById(Long id) {
-        final int isDeleted = stationDao.deleteById(id);
+        Set<Long> stationIds = new HashSet<>();
+        final List<Section> sections = sectionDao.findAll();
 
+        getStationIds(stationIds, sections);
+        validateDeletable(id, stationIds);
+
+        final int isDeleted = stationDao.deleteById(id);
         if (isDeleted == DELETE_FAIL) {
             throw new NotExistException("존재하지 않는 지하철 역입니다.");
+        }
+    }
+
+    private void getStationIds(Set<Long> stationIds, List<Section> sections) {
+        for (Section section : sections) {
+            final Long upStationId = section.getUpStationId();
+            final Long downStationId = section.getDownStationId();
+
+            stationIds.add(upStationId);
+            stationIds.add(downStationId);
+        }
+    }
+
+    private void validateDeletable(Long id, Set<Long> stationIds) {
+        if (stationIds.contains(id)) {
+            throw new IllegalArgumentException("구간을 먼저 삭제해주세요.");
         }
     }
 }
